LayoutTests/platform/ios-simulator-16-wk2/imported/w3c/web-platform-tests/css/css-tables/table-model-fixup-2-expected.txt

Table fixup

Check that replaced elements display types are handled properly in fixup

Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top)


Both text inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- img elements


Both text inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=text elements


Both button inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=button elements


Both file inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=file elements


Both images should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- img elements



PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top) assert_equals: expected 846.359375 but got 878.359375
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top) assert_equals: expected 1633.359375 but got 1665.359375
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top) assert_equals: expected 2410.359375 but got 2442.359375
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
PASS Replaced elements outside a table cannot be table and are considered block -- img elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements assert_equals: expected 3943.328125 but got 3963.328125
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- img elements assert_equals: expected 4085.328125 but got 4117.328125
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements assert_equals: expected 4501.328125 but got 4521.328125
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements assert_equals: expected 4643.328125 but got 4675.328125
PASS Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 5059.328125 but got 5079.328125
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- img elements assert_equals: expected 5201.328125 but got 5233.328125
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 5617.328125 but got 5637.328125
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- img elements assert_equals: expected 5759.328125 but got 5791.328125
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements assert_equals: expected 6175.328125 but got 6195.328125
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- img elements assert_equals: expected 6317.328125 but got 6349.328125
